Added suggestions

+++ shm/src/alcohol.cpp
@@ -7,5 +7,5 @@ Alcohol::Alcohol(const std::string& name, size_t amount, size_t basePrice, size_
     : Cargo(name, amount, basePrice), power_(power) {}
 
 size_t Alcohol::getPrice() const {
-    return static_cast<size_t>((float)basePrice_ * ((float)power_ / BASE_POWER));
+    return static_cast<size_t>((float)basePrice_ * (power_ / BASE_POWER));
 }
